# res/openapi/huki-ebook-openapi.generated.json
{
  "components": {
    "schemas": {
...
      "commerce_AddCartItemDto": {
        "properties": {
          "format": {
            "enum": [
              "PHYSICAL",
              "DIGITAL"
            ],
            "type": "string"
          }
        },
        "required": [
          "format"
        ],
        "type": "object"
      },
...
      "commerce_CheckoutConfirmDto": {
        "properties": {
          "paymentMethod": {
            "enum": [
              "COD",
              "ONLINE_PAYMENT"
            ],
            "type": "string"
          },
          "paymentProvider": {
            "type": "string"
          },
          "sessionId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "sessionId",
          "paymentMethod"
        ],
        "type": "object"
      },
      "commerce_CheckoutPreviewDto": {
        "properties": {
          "note": {
            "type": "string"
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/commerce_ShippingAddressDto"
          }
        },
        "type": "object"
      },
...
      "commerce_ShippingAddressDto": {
        "properties": {
          "district": {
            "type": "string"
          },
          "line1": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "recipientName": {
            "type": "string"
          },
          "ward": {
            "type": "string"
          }
        },
        "required": [
          "recipientName",
          "phone",
          "line1",
          "ward",
          "district",
          "province"
        ],
        "type": "object"
      },
...
      "commerce_UpdateCartItemDto": {
        "properties": {},
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "commerce_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "community_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "identity_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "JWT-auth": {
        "bearerFormat": "JWT",
        "description": "Enter JWT token",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      },
      "JWT-refresh": {
        "bearerFormat": "JWT",
        "description": "Enter JWT token",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      },
      "shipping_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "https://huki.ebook.com",
      "name": "HUKI Support",
      "url": "support@huki.ebook.com"
    },
    "description": "\n# HUKI EBOOK Platform API\n\nWelcome to the HUKI EBOOK API documentation. This API provides endpoints for managing books, orders, payments, and more.\n\n## Base URL\n`http://localhost:3000/api/v1`\n\n## Authentication\nMost endpoints require JWT Bearer token authentication.\n1. Register/Login to get access token\n2. Include header: `Authorization: Bearer <token>`\n\n## Services\n| Service | Port | Description |\n|---------|------|-------------|\n| Identity | 3001 | Auth, Users, Sessions |\n| Business | 3002 | Business, Stores, Members |\n| Commerce | 3003 | Books, Cart, Orders, Payments |\n| Shipping | 3004 | Shipments, Addresses |\n| Community | 3005 | Forum, Chat, Reviews |\n| Promotion | 3007 | Vouchers, Banners |\n\n## Common Headers\n| Header | Required | Description |\n|--------|----------|-------------|\n| Content-Type | Yes | application/json |\n| Authorization | No | Bearer token |\n| Idempotency-Key | For checkout | Unique key for idempotent operations |\n\n## Error Codes\n- 400: Bad Request\n- 401: Unauthorized\n- 403: Forbidden\n- 404: Not Found\n- 409: Conflict\n- 500: Server Error\n\n---\n*Last updated: 2026-08-24*\n    ",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://huki.ebook.com/terms",
...
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/cart": {
      "delete": {
        "description": "Removes all items from the user's cart.",
        "operationId": "CartController_clear",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cart cleared"
          },
          "401": {
            "description": "Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Clear entire cart",
        "tags": [
          "Cart"
        ]
      },
      "get": {
        "description": "Returns the current user's shopping cart with all items.",
        "operationId": "CartController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cart details"
          },
          "401": {
            "description": "Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user cart",
        "tags": [
          "Cart"
        ]
      }
    },
    "/api/v1/cart/checkout/confirm": {
      "post": {
        "description": "Confirms the checkout and creates the order. Requires Idempotency-Key header.",
        "operationId": "CheckoutController_confirm",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Unique key to prevent duplicate orders (UUID recommended)",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commerce_CheckoutConfirmDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Order created successfully"
          },
          "400": {
            "description": "Invalid checkout data or expired quote"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "404": {
            "description": "Cart, address, or item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm checkout",
        "tags": [
          "Checkout"
        ]
      }
    },
    "/api/v1/cart/checkout/preview": {
      "post": {
        "description": "Returns a preview of the checkout including shipping options, payment methods, and total.",
        "operationId": "CheckoutController_preview",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commerce_CheckoutPreviewDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Checkout preview"
          },
          "400": {
            "description": "Invalid checkout data"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "404": {
            "description": "Cart or item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview checkout",
        "tags": [
          "Checkout"
        ]
      }
    },
    "/api/v1/cart/items": {
      "post": {
        "description": "Adds a book to the user's shopping cart.",
        "operationId": "CartController_add",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commerce_AddCartItemDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Item added to cart"
          },
          "400": {
            "description": "Invalid item or quantity"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "404": {
            "description": "Book not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add item to cart",
        "tags": [
          "Cart"
        ]
      }
    },
    "/api/v1/cart/items/{itemId}": {
      "delete": {
        "description": "Removes a specific item from the cart.",
        "operationId": "CartController_remove",
        "parameters": [
          {
            "description": "Cart item ID",
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item removed from cart"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "404": {
            "description": "Cart item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove item from cart",
        "tags": [
          "Cart"
        ]
      },
      "patch": {
        "description": "Updates the quantity of an item in the cart.",
        "operationId": "CartController_update",
        "parameters": [
          {
            "description": "Cart item ID",
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commerce_UpdateCartItemDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Item quantity updated"
          },
          "400": {
            "description": "Invalid quantity"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "404": {
            "description": "Cart item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update cart item quantity",
        "tags": [
          "Cart"
        ]
      }
    },
...
    "/api/v1/seller/orders/{id}/confirm": {
      "patch": {
        "operationId": "SellerOrdersController_confirm",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm a new order",
        "tags": [
          "Seller orders"
        ]
      }
    },
...
  },
  "servers": [
...
}